requirements.txt
streamlit==1.50.0
plotly==6.3.1
python-dotenv==1.1.1
supabase==2.16.0
plotly==6.3.1
streamlit-url-fragments==0.1.2